Moves a test file to the testutils package and updates references accordingly
