+++ tests/com/tkomiya/tests/VocabListJsonConverterTest.java
@@ -13,6 +13,7 @@ import com.tkomiya.infrastructure.FileUtilities;
 import com.tkomiya.infrastructure.VocabListJsonConverter;
 import com.tkomiya.models.VocabList;
 import com.tkomiya.testutils.Constants;
+import com.tkomiya.testutils.TestVocabListLoader;
 
 public class VocabListJsonConverterTest {
